+++ game/combat/aihaywire.py
@@ -1,35 +1,42 @@
 import pbge
 import random
 import gears
+from . import actions
 
 
 class HaywireTurn(object):
     def __init__(self,pc: gears.base.Being,camp):
         self.pc = pc
+        self.current_pos = self.pc.pos
         self.camp = camp
-        self.act()
+        self.mp = self.camp.fight.cstat[self.pc].mp_remaining
+        self.ap = self.camp.fight.cstat[self.pc].action_points
 
-    def bust_a_move(self):
+    def bust_a_move(self) -> list:
+        my_actions = list()
+
+        my_actions.append(actions.Faff(self.camp, self.pc, gears.geffects.HaywireAnim))
+        self.ap -= 1
+    
         # Move to a randomly selected accessible tile.
-        if not hasattr(self.pc,"mmode"):
-            self.camp.fight.cstat[self.pc].spend_ap(1)
-            return
-        elif not self.pc.mmode:
-            self.pc.gear_up(self.camp.scene)
-        if self.pc.get_current_speed() > 10:
-            mynav = pbge.scenes.pathfinding.NavigationGuide(self.camp.scene, self.pc.pos, self.pc.get_current_speed(),
+        if self.mp > 10:
+            mynav = pbge.scenes.pathfinding.NavigationGuide(self.camp.scene, self.current_pos, self.mp,
                                                             self.pc.mmode, self.camp.scene.get_blocked_tiles())
             mydest = random.choice(list(mynav.cost_to_tile.keys()))
-            self.camp.fight.move_model_to(self.pc, mynav, mydest)
-        else:
-            self.camp.fight.cstat[self.pc].spend_ap(1)
 
-    def act(self):
+            self.mp -= mynav.cost_to_tile[mydest]
+            self.current_pos = mydest
+            my_actions.append(actions.MoveModelToPos(self.camp, self.pc, mynav, mydest))
+
+        return my_actions
+
+    def get_actions(self) -> list:
+        my_actions = list()
         if random.randint(1,5) != 1:
-            pbge.my_state.view.play_anims(gears.geffects.HaywireAnim(pos=self.pc.pos))
-            self.bust_a_move()
+            my_actions += self.bust_a_move()
+            while self.ap > 0 and random.randint(1,5) == 2:
+                my_actions += self.bust_a_move()
 
-            while self.camp.fight.still_fighting() and self.camp.fight.cstat[self.pc].action_points > 0 and random.randint(1,3) == 1:
-                self.bust_a_move()
+        return my_actions