+++ gears/usables.py
@@ -46,6 +46,28 @@ class QuickFixPill(Singleton):
         return mylist
 
 
+class MiracleCure(Singleton):
+    VALUE = 100000
+    @classmethod
+    def get_invocations(cls, _pc):
+        mylist = list()
+        mylist.append(pbge.effects.Invocation(
+            name = 'Quick Fix',
+            fx=geffects.DoHealing(
+                10,6,repair_type=materials.RT_MEDICINE,
+                anim = geffects.MedicineAnim,
+                ),
+            area=pbge.scenes.targetarea.SelfOnly(),
+            used_in_combat = True, used_in_exploration=True,
+            ai_tar = aitargeters.GenericTargeter(impulse_score=50,conditions=[aitargeters.TargetIsAlly(),aitargeters.TargetIsOperational(),aitargeters.TargetIsDamaged(materials.RT_MEDICINE)],targetable_types=pbge.scenes.PlaceableThing),
+            shot_anim=None,
+            data=geffects.AttackData(pbge.image.Image('sys_skillicons.png',32,32),0),
+            price=[],
+            targets=1
+        ))
+        return mylist
+
+
 class DuctTape(Singleton):
     VALUE = 75
     @classmethod